test backup

+++ src/main/java/com/enigma/walletkurs/models/TradingEntity.java
@@ -4,8 +4,6 @@ import java.sql.Date;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToOne;
@@ -38,10 +36,10 @@ public class TradingEntity {
 	}
 
 	@Column(name="amount")
-	Float amount;
+	Double amount;
 	
 	@Column(name="income")
-	Float income;
+	Double income;
 
 	public String getTradingId() {
 		return tradingId;
@@ -80,21 +78,20 @@ public class TradingEntity {
 		return rateId;
 	}
 
-	public Float getAmount() {
+	public Double getAmount() {
 		return amount;
 	}
 
-	public void setAmount(Float amount) {
+	public void setAmount(Double amount) {
 		this.amount = amount;
 	}
 
-	public Float getIncome() {
+	public Double getIncome() {
 		return income;
 	}
 
-	public void setIncome(Float income) {
+	public void setIncome(Double income) {
 		this.income = income;
 	}
-	
-	
+
 }
